+++ clients/src/test/java/org/apache/kafka/clients/NetworkClientTest.java
@@ -52,7 +52,7 @@ public class NetworkClientTest {
     private int nodeId = 1;
     private Cluster cluster = TestUtils.singletonCluster("test", nodeId);
     private Node node = cluster.nodes().get(0);
-    private long reconnectBackoffMsTest = 10 * 1000;
+    private int reconnectBackoffMsTest = 10 * 1000;
     private NetworkClient client = new NetworkClient(selector, metadata, "mock", Integer.MAX_VALUE, reconnectBackoffMsTest, 
             64 * 1024, 64 * 1024, requestTimeoutMs, time);
