+++ clients/src/test/java/org/apache/kafka/clients/consumer/KafkaConsumerTest.java
@@ -1096,8 +1096,8 @@ public class KafkaConsumerTest {
         String clientId = "mock-consumer";
         String groupId = "mock-group";
         String metricGroupPrefix = "consumer";
-        long retryBackoffMs = 100;
-        long requestTimeoutMs = 30000;
+        int retryBackoffMs = 100;
+        int requestTimeoutMs = 30000;
         boolean excludeInternalTopics = true;
         int minBytes = 1;
         int maxBytes = Integer.MAX_VALUE;
